Unify remote menu and start entry items for desktop
